@@ -191,7 +191,7 @@ public class DlnaDeviceProfile : DeviceProfile
         return TranscodingProfiles
             .Where(i => i.Type == DlnaProfileType.Audio)
             .Where(i => string.Equals(container, i.Container, StringComparison.OrdinalIgnoreCase))
-            .FirstOrDefault(i => ContainerHelper.Split(i.AudioCodec).Contains(audioCodec ?? string.Empty, StringComparison.OrdinalIgnoreCase));
+            .FirstOrDefault(i => ContainerHelper.ContainsContainer(i.AudioCodec, audioCodec));
     }
 
     /// <summary>
@@ -208,7 +208,7 @@ public class DlnaDeviceProfile : DeviceProfile
         return TranscodingProfiles
             .Where(i => i.Type == DlnaProfileType.Video)
             .Where(i => string.Equals(container, i.Container, StringComparison.OrdinalIgnoreCase))
-            .Where(i => ContainerHelper.Split(i.AudioCodec).Contains(audioCodec ?? string.Empty, StringComparison.OrdinalIgnoreCase))
+            .Where(i => ContainerHelper.ContainsContainer(i.AudioCodec, audioCodec))
             .FirstOrDefault(i => string.Equals(videoCodec, i.VideoCodec, StringComparison.OrdinalIgnoreCase));
     }
 
@@ -331,6 +331,7 @@ public class DlnaDeviceProfile : DeviceProfile
     /// <param name="refFrames">The ref frames.</param>
     /// <param name="numVideoStreams">The number of video streams.</param>
     /// <param name="numAudioStreams">The number of audio streams.</param>
+    /// <param name="numStreams">The number of streams.</param>
     /// <param name="videoCodecTag">The video Codec tag.</param>
     /// <param name="isAvc">True if Avc.</param>
     /// <returns>The <see cref="ResponseProfile"/>.</returns>
@@ -353,6 +354,7 @@ public class DlnaDeviceProfile : DeviceProfile
         int? refFrames,
         int? numVideoStreams,
         int? numAudioStreams,
+        int numStreams,
         string? videoCodecTag,
         bool? isAvc)
     {
@@ -383,7 +385,26 @@ public class DlnaDeviceProfile : DeviceProfile
             var anyOff = false;
             foreach (ProfileCondition c in i.Conditions)
             {
-                if (!ConditionProcessor.IsVideoConditionSatisfied(GetModelProfileCondition(c), width, height, bitDepth, videoBitrate, videoProfile, videoRangeType, videoLevel, videoFramerate, packetLength, timestamp, isAnamorphic, isInterlaced, refFrames, numVideoStreams, numAudioStreams, videoCodecTag, isAvc))
+                if (!ConditionProcessor.IsVideoConditionSatisfied(
+                        GetModelProfileCondition(c),
+                        width,
+                        height,
+                        bitDepth,
+                        videoBitrate,
+                        videoProfile,
+                        videoRangeType,
+                        videoLevel,
+                        videoFramerate,
+                        packetLength,
+                        timestamp,
+                        isAnamorphic,
+                        isInterlaced,
+                        refFrames,
+                        numStreams,
+                        numVideoStreams,
+                        numAudioStreams,
+                        videoCodecTag,
+                        isAvc))
                 {
                     anyOff = true;
                     break;
